RDriver's base is complete
Yup, its done. 
There is some room of course to move forward with this. For starters, we need to anticipate output from various graphical packages so we can return their output accordingly. 
This leads into the point of ensuring that all libraries and packages behave correctly and we need to do this by testing them against our driver.
However, the base RDriver is complete now and we can move forward to connecting it to the server now