More testing and refining of our R-driver
Did more testing and refining of the driver. It is able to import a plethora of different libraries and I have tested many of the more popular libraries against it. So far, so good. Though we may need to think about diversifying our exports for the website to account for R's more diverse usage...