Implemented protection for users and minor changes
Using a new technology: docker, we now have protection from other user's code in our project. Also changed the paths to be user agnostic and the repo is now placed in the bin folder in basic linux distros. Also removed R's path in our program as R no longer is necessary for a user's computer